Refactor SaveActivityInteractor

+++ app/src/main/java/com/gpetuhov/android/hive/domain/repository/Repo.kt
@@ -60,7 +60,7 @@ interface Repo {
     fun saveUserWork(newWork: String, onError: () -> Unit)
     fun saveUserInterests(newInterests: String, onError: () -> Unit)
     fun saveUserStatus(newStatus: String, onError: () -> Unit)
-    fun saveUserActivity(newActivity: Int)
+    fun saveUserActivity(newActivity: Long)
     fun setUserOnline()
     fun setUserOffline()
     fun deleteUserDataRemote(onSuccess: () -> Unit, onError: () -> Unit)
